@alxbilger 2023-05-05
Hi @fredroy,
It seems this PR can cause a crash, at least on MacOS. The crash happens in `OglModel.cpp`, line 875 (`glMapBufferRange`).