@fredroy 2023-05-05
Crap, I suppose this is because macos cannot run in compat mode (aka mixing ogl legacy (v2) and ogl core (>v3) , as it is opengl>3 functions https://registry.khronos.org/OpenGL-Refpages/gl4/html/glMapBufferRange.xhtml

2 choices:
 - ugly ifdef
 - revert the change and call glBufferData instead()